chore: update email package #672
- wow, resend is awesome
- moved components to email package
- updated props and params so things work smoove
- removed packages from ai app
- also minor edits to utils file and contributing

Closes #672
