+++ apps/ai/package.json
@@ -35,7 +35,6 @@
     "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-popover": "^1.0.7",
     "@radix-ui/react-slot": "^1.0.2",
-    "@react-email/components": "^0.0.14",
     "@rive-app/react-canvas-lite": "^4.7.1",
     "@svgr/webpack": "^8.1.0",
     "@t3-oss/env-nextjs": "^0.7.3",
@@ -108,7 +107,6 @@
     "react-colorful": "^5.6.1",
     "react-confetti": "^6.1.0",
     "react-dom": "^18.2.0",
-    "react-email": "^2.0.0",
     "react-hook-form": "^7.49.3",
     "react-hot-toast": "^2.4.1",
     "react-wrap-balancer": "^1.1.0",
