+++ apps/ai/src/lib/const.ts
@@ -1,4 +1,4 @@
-const VERCEL_URL = process.env.NEXT_PUBLIC_VERCEL_URL
+export const VERCEL_URL = process.env.NEXT_PUBLIC_VERCEL_URL
   ? `https://${process.env.NEXT_PUBLIC_VERCEL_URL}`
   : '';
