+++ packages/email/package.json
@@ -4,12 +4,17 @@
   "version": "0.1.0",
   "type": "module",
   "exports": {
-    ".": "./src/index.ts"
+    ".": "./src/index.ts",
+    "./abuse-report-email": "./src/emails/abuse-report-email.tsx",
+    "./magic-link-email": "./src/emails/magic-link-email.tsx",
+    "./team-invite-email": "./src/emails/team-invite-email.tsx",
+    "./welcome-email": "./src/emails/welcome-email.tsx"
   },
   "license": "MIT",
   "scripts": {
     "clean": "rm -rf .turbo node_modules",
     "format": "prettier --check . --ignore-path ../../.gitignore",
+    "email": "email dev --dir ./src/emails --port 3333",
     "knip": "knip",
     "lint": "eslint .",
     "typecheck": "tsc --noEmit"
@@ -18,6 +23,8 @@
     "@and-voila/eslint-config": "workspace:^0.2.0",
     "@and-voila/prettier-config": "workspace:^0.1.0",
     "@and-voila/tsconfig": "workspace:^0.1.0",
+    "@types/react": "^18.2.48",
+    "@types/react-dom": "^18.2.18",
     "eslint": "^8.56.0",
     "knip": "^4.1.0",
     "prettier": "^3.2.2",
@@ -30,6 +37,9 @@
   },
   "prettier": "@and-voila/prettier-config",
   "dependencies": {
+    "@and-voila/utils": "workspace:^0.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-email": "^2.0.0",
     "@react-email/components": "^0.0.14"
   }
