+++ packages/email/src/emails/abuse-report-email.tsx
@@ -12,26 +12,27 @@ import {
   Text,
 } from '@react-email/components';
 
-import { SITE_URL } from '#/lib/const';
+const subject = '😱 Abuse Report Notification';
 
-interface MagicLinkEmailProps {
-  actionUrl: string;
-  mailType: 'login' | 'register';
+export interface SendAbuseReportEmailProps {
+  urlToReport: string;
   siteName: string;
+  siteUrl: string;
 }
 
-export const MagicLinkEmail = ({
-  actionUrl,
-  mailType,
-}: MagicLinkEmailProps) => (
+export const SendAbuseReportEmail = ({
+  urlToReport,
+  siteName,
+  siteUrl,
+}: SendAbuseReportEmailProps) => (
   <Html>
     <Head />
-    <Preview>Here&apos;s the magic link you requested from And Voila.</Preview>
+    <Preview>{subject}</Preview>
     <Tailwind>
       <Body className="bg-[#B4B4B4] py-[24px] font-sans">
         <Container className="mx-auto rounded-lg bg-[#F8F9FB] px-[24px] pt-[36px]">
           <Img
-            src={`${SITE_URL}/and-voila-logo.png`}
+            src={`${siteUrl}/and-voila-logo.png`}
             width="150"
             height="34"
             alt="A colored logomark with an ampersand and exclamation icon and the the Voila"
@@ -39,36 +40,28 @@ export const MagicLinkEmail = ({
           />
           <Hr className="border-t-1 mt-[16px] border-[#d20f9a]" />
           <Text className="mt-[40px] text-[22px] leading-[28px] text-[#6E6E6E]">
-            🌺 Aloha,
+            Head&apos;s up!
           </Text>
-          {mailType === 'login' ? (
-            <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
-              We&apos;re excited to have you back. Click the button below to log
-              in.
-            </Text>
-          ) : (
-            <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
-              We&apos;re thrilled to welcome you aboard. Just one more click and
-              you&apos;re in. Activate your account with the button below.
-            </Text>
-          )}
+          <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
+            We have received an abuse report for: {siteName}
+          </Text>
+          <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
+            The reported URL is: {urlToReport}
+          </Text>
+          <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
+            Review the content and take appropriate action as per our Terms of
+            Service.
+          </Text>
           <Section className="mt-[48px] text-center">
             <Button
               className="inline-block w-[85%] rounded-md bg-[#18A300] px-[32px] py-[16px] text-[22px] leading-[28px] text-[#F8F9FB] no-underline"
-              href={actionUrl}
+              href={urlToReport}
             >
-              {mailType === 'login' ? 'Log In' : 'Activate Account'}
+              View Reported URL
             </Button>
           </Section>
-          <Text className="text-center text-[12px] text-[#6E6E6E]">
-            This link will expire in 24 hours and is for one-time use only.
-          </Text>
           <Hr className="border-t-1 mt-[48px] border-[#d20f9a]" />
           <Text className="mt-[40px] text-[12px] text-[#6E6E6E]">
-            If you didn&apos;t request this magic link, feel free to disregard
-            this email.
-          </Text>
-          <Text className="text-[12px]">
             <span className="font-bold">BRIL.LA, LLC.</span> 1370 N. St. Andrews
             Place, Los Angeles, CA 90028
           </Text>
@@ -78,4 +71,4 @@ export const MagicLinkEmail = ({
   </Html>
 );
 
-export default MagicLinkEmail;
+export default SendAbuseReportEmail;
