+++ packages/email/src/emails/magic-link-email.tsx
@@ -12,29 +12,26 @@ import {
   Text,
 } from '@react-email/components';
 
-import { SITE_URL } from '#/lib/const';
-
-export const TemplateName = 'send-abuse-report-email';
-
-const subject = '😱 Abuse Report Notification';
-
-export interface SendAbuseReportEmailProps {
-  urlToReport: string;
+interface MagicLinkEmailProps {
+  actionUrl: string;
+  mailType: 'login' | 'register';
   siteName: string;
+  siteUrl: string;
 }
 
-export const SendAbuseReportEmail = ({
-  urlToReport,
-  siteName,
-}: SendAbuseReportEmailProps) => (
+export const MagicLinkEmail = ({
+  actionUrl,
+  mailType,
+  siteUrl,
+}: MagicLinkEmailProps) => (
   <Html>
     <Head />
-    <Preview>{subject}</Preview>
+    <Preview>Here&apos;s the magic link you requested from And Voila.</Preview>
     <Tailwind>
       <Body className="bg-[#B4B4B4] py-[24px] font-sans">
         <Container className="mx-auto rounded-lg bg-[#F8F9FB] px-[24px] pt-[36px]">
           <Img
-            src={`${SITE_URL}/and-voila-logo.png`}
+            src={`${siteUrl}/and-voila-logo.png`}
             width="150"
             height="34"
             alt="A colored logomark with an ampersand and exclamation icon and the the Voila"
@@ -42,28 +39,36 @@ export const SendAbuseReportEmail = ({
           />
           <Hr className="border-t-1 mt-[16px] border-[#d20f9a]" />
           <Text className="mt-[40px] text-[22px] leading-[28px] text-[#6E6E6E]">
-            Head&apos;s up!
+            🌺 Aloha,
           </Text>
-          <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
-            We have received an abuse report for: {siteName}
-          </Text>
-          <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
-            The reported URL is: {urlToReport}
-          </Text>
-          <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
-            Review the content and take appropriate action as per our Terms of
-            Service.
-          </Text>
+          {mailType === 'login' ? (
+            <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
+              We&apos;re excited to have you back. Click the button below to log
+              in.
+            </Text>
+          ) : (
+            <Text className="text-[22px] leading-[28px] text-[#6E6E6E]">
+              We&apos;re thrilled to welcome you aboard. Just one more click and
+              you&apos;re in. Activate your account with the button below.
+            </Text>
+          )}
           <Section className="mt-[48px] text-center">
             <Button
               className="inline-block w-[85%] rounded-md bg-[#18A300] px-[32px] py-[16px] text-[22px] leading-[28px] text-[#F8F9FB] no-underline"
-              href={urlToReport}
+              href={actionUrl}
             >
-              View Reported URL
+              {mailType === 'login' ? 'Log In' : 'Activate Account'}
             </Button>
           </Section>
+          <Text className="text-center text-[12px] text-[#6E6E6E]">
+            This link will expire in 24 hours and is for one-time use only.
+          </Text>
           <Hr className="border-t-1 mt-[48px] border-[#d20f9a]" />
           <Text className="mt-[40px] text-[12px] text-[#6E6E6E]">
+            If you didn&apos;t request this magic link, feel free to disregard
+            this email.
+          </Text>
+          <Text className="text-[12px]">
             <span className="font-bold">BRIL.LA, LLC.</span> 1370 N. St. Andrews
             Place, Los Angeles, CA 90028
           </Text>
@@ -72,3 +77,5 @@ export const SendAbuseReportEmail = ({
     </Tailwind>
   </Html>
 );
+
+export default MagicLinkEmail;
